src/install: properly handle how slots are marked active/inactive
When performing an installation, we need to deactivate booting the
relevant set of slots before actually updating them, and after a
successful installation, we must (by default) mark the set of slots as
active (primary) again.

Activating and deactivating here means that we tell the boot selection
layer (and thus the underlying bootloader used) to switch order,
priority, attempts counter, etc. in order to assure the selected slot is
not booted anymore (deactivated) or booted as the primary one (activated).

As a bootloader only boots exactly *one single* boot target,
a target group, as computed and used by RAUC, must contain exactly *one*
slot that serves as the interface to the bootloader, i.e. that has a
'bootname' set.

We can neither atomically install if we cannot deactivate the slot
group, nor we can (atomically) install to slot groups where multiple slots
are activatable by the bootloader.

The current implementation of marking a slot 'bootable'/'non-bootable'
simply iterates over a loop of all 'install_images' and calls the
mark function for each slot having a bootname and no parent set.

Despite this works in most cases, this is wrong on several levels:

The loop iteration potentially allows that

  a) when we have no bootname set or update no slot with a bootname set,
  silently no slot will be activated (similar to what we have above).

  b) when we have multiple slots with a bootname set (misconfiguration),
  then the mark_active() call will silently be executed for all of them,
  which is as wrong as not executing it.

We solve this all at once now by introducing the function
get_bootname_slot():

It returns *exactly one* slot from a target install group that has a
'bootname' set.

For cases where none or multiple slots have set a bootname, the function
returns with a GError set.

Then we simply call `mark_active()`/`r_boot_set_state()` on the
(single) slot returned.

To not break setups that misused this feature for now, we emit only a
warning if get_bootname_slot() failed instead of aborting.

Future versions should handle this as a critical error instead!

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>