src/config_file: add support for getting variant over executable
This introduces the new system config key 'variant-bin' which acts
similar to the existing 'variant-name', 'variant-file', etc.

With 'variant-bin' you can specify the path to a script or binary
(executable) that prints out the variant of the board.

A possible use-case is to have the variant information encoded in
hardware bits and an application that may read and transform these to a
human-readable string.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>