git-version-gen: drop rc detection and improve detection of old git instead
This undoes a workaround that was introduced for v1.0-rc1 to be able to
create a release tarball with halfway sane version string. By making the
glob expressions stricter a clean -rc release isn't matched and so it is
not tried to add a separate numeric component for the count of commits
since the last tag.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>