src/slot: normalize device names to find mounted slots
This way a mounted slot can be found if the path used to mount the
device is not the same as the path in the rauc configuration file.
E.g., the configuration file uses a link like
/dev/disk/by-partlabel/rootfs0 while the kernel's list of mounted
devices uses a name like /dev/mmcblk1p2.

This is a fallback after simply comparing the paths, as if the paths
are the same then the devices must the same too.  This avoids changing
behavior in cases there the devices do not exist, for instance if rauc
status before a device is online.

Not being able to detect that the current rootfs is already mounted
causes rauc to try to mount the slot, which will fail.

Fixes #388.

Signed-off-by: Trent Piepho <tpiepho@impinj.com>