test/config_file: refactor simple typo tests
There's a lot of common boilerplate in the typo tests; the only
essential varying thing is the actual contents of the config file.

The temporary filename used is now standardized as typo.conf, but that
seems better than the copy-pasted invalid_bootloader.conf - in one
case, the more explicit invalid_max_bundle_download_size.conf was
used, but it didn't seem worth it to let that be a parameter the
config_file_typo() helper.

Also, only some of the typo tests had a g_assert_null(config); that
might as well be done for all of them.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>
Signed-off-by: Jan Luebbe <jlu@pengutronix.de>