src/main: show slot name also in 'Booted from' line
This eases comparing the currently booted slot with the activated as one
can now simply compare the slot name.

Comparing is quite complicated in use cases where the 'bootslot' value
is not derived from the bootname but form the booted device. This is
true for all cases where the bootloader does not or cannot alter the
kernel commandline to provide 'rauc.slot' or 'bootchooser.active'.

In order to get the right slot referred by the 'bootname', we also need
some more slot search utility functions here.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>