src/main: do not print status dump when calling mark-* subcommand
Prior to this e.g. a call of

  rauc status mark-active other

always dumped the status output as one would expect when calling 'rauc
status'. As this printout occured *before* actually applying the
requested mark-* action, the output shows the old status instead of the
new on.

This is quite misleading.

Anyway, there is no need to always dump the status when modifying it. We
already get a message telling us what was actually changed.

Thus we simply disable calling rauc status when invoking a sub-command.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>