src/utils.c: only return newly-allocated objects in resolve_path()
resolve_path() returned a newly allocated gchar when a relative path was
provided as 'path' argument while it returned the pointer to path itself
when 'path' already was an absolute path.

This will break all assumption callers of resolve_path() can make on if
and when to free the returned gchar.

Fixed this by always returning a newly allocated gchar.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>