Fix memory leaks in D-Bus notification callbacks (refs #426)
According to glib manual, g_variant_lookup() is a wrapper using g_variant_get()
internally, so the same GVariant format strings rules apply.
A format string "s" returns already a copy of the string while "&s" borrows
the memory. In our case, we pass the received pointers to g_strdup_*() functions
and thus we care about copying data ourself. So using "&s" instead
prevents leaking memory.

And on the second/receiving side, we pop the pointer from the queue and
have to free it now.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>