Fix memleaks in resolve_bundle_path (refs #426)
1)
According to documentation g_get_current_dir() returned string should
be freed when no longer needed. So use a g_autofree decorated variable
to buffer the returned pointer.

2)
g_filename_from_uri() allocs memory for the hostname component if found.
But in this case (hostname != NULL) we leave the funtion with return
which does not free (the now assigned) hostname.
We can use g_autofree decoration to fix this without the need to care
about explicitely.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>