src/update_handler: add support for updating boot partitions in MBR
This introduces the possibility to update the bootloader of systems which
boot from a partition in the MBR partition table.
To differentiate between active and inactive partition, two new
configurations are introduced for slots: region-start and region-end.

After the inactive partition is written, the partition is activated by
changing the partition entry and rewriting the MBR.

Signed-off-by: Thomas Haemmerle <thomas.haemmerle@wolfvision.net>