src/install: replace 'force-install-same' key by 'install-same'
This replaces the former 'force-install-same' key by a new
'install-same' one to reflect that we intend to invert the default.

The new key is neutral concerning any default and will hopefully prevent
us from further renaming in the future.

As usual, the old key will rename valid to stay compatible with old
configs.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>