src/config_file: invert slot skipping logic
This simple change has a major impact of installation handling.

From now on, the default behavior of RAUC will be to always install an
image, regardless of whether its target slot checksum is the same as the
image's checksum.

This inverts the default slot skipping optimization present since the
initial release of RAUC.

Explicitly enabling this handling is still possible of course, by simply
specifying

  [slot.rootfs.0]
  ...
  install-same=false

in your system config.

This change will not have any impact on compatibility with old setup
despite that you do not gain minor speed performance in rare cases by
default.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>