src/main: a more fancy printout
So far, especially when having more complex slot setups, the output of
'rauc status' did not provide an easy and clear view of the slots
constellation and their respective status (booted, active, inactive).

This patch aims to fix this by giving a more tidied-up view with colors
and bold font for highlighting.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>