src/install: fix detection of unmounted slots in determine_slot_states()
By iterating over the list of mountpoints and setting a slots
ext_mount_point property in case a valid mount is detected is perfectly
fine for detecting mounts that have not been present before and are now.

But, if we run again and a mount that was previously there now
disappeared, we will come along and update the ext_mount_point property
anymore despite we should remove the mount point information then.

This is now fixed by clearing all slots ext_mount_point properties
before re-iterating over the mount point list. This will leave slots
that do not have an external mount anymore untouched and thus having a
NULL ext_mount_point as it should be.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>