src/utils: add subprocess wrapper function r_subprocess_new()
This aims to fulfill the same purpose that the newv variant
r_subprocess_newv() does, i.e. it adds a wrapper for logging these calls
that were not visible in subprocess debug log so far.
However, we need a little more code to extract and use the arguments.

The way we do it here is basically inspired by the original
implementation of g_subprocess_new().

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>