test/bundle: fix setting of certpath and keypath in resign test
The test used 'r_context_conf()->certpath' to set the certpath to a
different one than the fixture does. While using r_context_conf() is the
only default way to get a 'writable' non-const pointer to the context,
calling this function has another impact: It re-reads the entire context
and resets the conext 'pending' flag.
One result of this is that also the 'config->keyring_path' property set
in the test before will be overridden with its default. This leads to
using the wrong keyring (i.e. dev-ca.pem instead of dev-only-ca.pem).

However, Using the dev-ca keyring here will NOT make the test fail. But
as the dev-ca keyring validates both against a 'dev'-signed as well as
the 'rel'-signed bundle (and that's the sneaky part) the test gives a
false negative when resigning is broken in a way that was fixed by XXXX.

We fix this by accessing the normally module-hidden 'context' object
directly by declaring it as 'extern' in bundle.c. This way we assure the
context is not re-initialized and the more strict dev-only keyring is
used for checking now.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>