src/update_handler: allow casync to extract into redirected STDOUT
If a casync block-indexed image ("*.caibx") shall be installed to a UBI volume
casync has to extract to the volume's char device. This needs to start with rauc
ioctl'ing the file descriptor of the latter to receive an update and then
passing the prepared file descriptor as STDOUT to the spawned casync process.
Hence, we need to extend the related functions accordingly.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>