src/update_handler: filter out character devices as casync seeds
If extracting casync block-indexed images ("*.caibx") to a slot it could be of
advantage to use the slot's active counterpart as a casync seeding device. In
the case of UBI volume based slots those are char devices which can not be
handed out to casync. Instead one needs to derive related UBIBLOCK devices for
that purpose which have to be cleaned up afterwards. For the moment to simplify
things just discard the usage of seeding devices in that case.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>