src: introduce new bundle/resign/convert option '--signing-keyring'
When doing post-signing verification for 'rauc resign' we must be aware
that we need two different keyrings:

a) The one to verify the *original* bundle against
b) The one to verify the *resigned* bundle against

For now we only had the option `--keyring` and the same keyring was used
for both case a) and b).
While there are cases where this is valid (if the keyring accepts both
signatures), the default is that this will fail.

The existing '--keyring' parameter is already used for passing the
systems keyring to the RAUC service etc.
Thus for having a different keyring in post-signing verification we must
introduce another parameter to explicitly pass a different 'second' keyring.
This is what we do here with the '--signing-keyring' parameter.

Note that the '--keyring' parameter will still set both the bundle
verification and the post signing verification keyring.

To be consistent, this option is added for subcommands 'bundle',
'resign', and 'convert'.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>