src/signature.c: pass store to cms_verify instead of creating it there
This is a preparation to allow verification with different keyrings,
which is required for example in case of resigning.

With passing the store form outside the function it is possible to
previously set up stores for different CA locations.

This also allows to remove almost all context handling form the
corresponding tests and pass keyring store directly.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>