src/update_handler: fix emmc device detection
The current implementation assumes that the device path is always
something like /dev/mmcblk[0-9] but those paths are not deterministic.
Series [1] wanted to address this but never landed into mainline. So we
need to use a other mechanism and luckily there are the /dev/disk/by-*
paths. Now the problems are starting:
  - pre_install_checks() checks if the device exists and
  - img_to_boot_emmc_handler() needs the root device to decide the
active slot and then the handler concatenates the path by boot[0,1].
So the the /dev/disk/by-* won't work because I have either to specify:

[slot.bootloader.0]
device=/dev/disk/by-path/platform-219c000.usdhc
type=boot-emmc

This won't work because the path is
/dev/disk/by-path/platform-219c000.usdhc-boot[0,1]

or

[slot.bootloader.0]
device=/dev/disk/by-path/platform-219c000.usdhc-
type=boot-emmc

This won't work because of pre_install_checks() and
img_to_boot_emmc_handler() needs the root device node.

The fix is to resolve the symlink by using realpath. For more
information check 'man 3 realpath'.

[1] https://lore.kernel.org/patchwork/cover/674381/

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>