utils: add libc realpath wrapper
Most of the code base is written using glib and only a few times the
native libc functions are used. Since the glib don't provide a
g_realpath() helper we need to add a wrapper to for the libc realpath.
The wrapper is also a bit more user friendly because the caller only
have to specify the path which should be resolved.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>