remove misleading reference to ext2 / ext3 support.
Various places of the RAUC documentation noted ext2 and ext3 file system
support. This was quite misleading as RAUC has an ext4 file type, but
not an ext2 / ext3 type.

This removes the notions of ext2 and ext3 and adds a section in the FAQ
why there is no extra slot type available.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>