src/main: do not free propagated error in retrieve_status_via_dbus()
The glib error handling is designed for either handling or propagating
errors. With calling g_propagate_prefixed_error() and g_error_free() we
do both at once which violates this and results in a double-free because
g_propagate_prefixed_error() already cares for freeing the source error.

Fix this by not freeing the already propagated error.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>