src/main: do not prepend duplicate scope information in status error messages
For errors in retrieve_status_via_dbus() the string "rauc status: " was
prepended to the error message. Now when this error was printed as
result of 'status_start' function another "rauc mark: " string was
prepended.

This is something between redundant and confusing as it gives two but
different scope informations. As 'mark' commands are a subset of 'status
anyway, we change this to prepend 'rauc status: ' in the end and remove
the prior prepend completely.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>