src/network: limit file download size by using CURLOPT_MAXFILESIZE
Limiting the download file size is implemented in RAUC using libcurl
write and transfer callback functions that only abort if the maximum
download size is exceeded.
The downside of this is that in case of too large bundles, the resulting
error message is quite unspecific:

| Transfer failed: Failed writing body (0 != 1448)

When using curl's built-in CURLOPT_MAXFILESIZE instead, the error
message for too large downloads changes to:

| Transfer failed: Maximum file size exceeded

However, this only works for cases where the server announces the size
of the artifact prior to download. For other cases (possible both with
FTP and HTTP) we still have to rely on the abort through the callback
functions.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>