src/network: forward error message from write and transfer callback
In addition to the prior commit this allows to gain a more targeted
error message in cases where the server did not advertise the download
size before downloading.

Having the libcurl abort is still usefull as it allows to abort before
actually starting the download while this technique here only allows to
abort *during* the download process.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>