src/network: fix download size limitation on 32bit systems
Both download_file and download_mem only supported download limit
arguments of type size_t. This type is limited to the maximum size to
represent objects and thus only 32bit on 32bit systems which is
insufficient for larger downloads > 4GB.

As the limit value is mainly meant to be passed as argument to
curl_easy_setopt(), we use the curl types curl_off_t here instead of
goffset or off_t.

Note:
  Casting both sides of the transfer limit '>' comparison is required as
  on 64 bit systems the direct comparison results in

  | different signedness: ‘long unsigned int’ and ‘curl_off_t’ {aka ‘long int’}

  requiring to cast the right side to (guint64) while on 32 bit systems
  you then get

  | different signedness: 'off_t' {aka ‘long long int’} and ‘long long usigned int’

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>