Makefile.am: build with -f{function,data}-sections and -Wl,--gc-sections
Make the generated binary smaller by having the compiler and linker
work together to eliminate functions that are not used in a given
configuration. The same can theoretically be achieved with proper use
of #ifdefs everywhere, but it is much more flexible and convenient to
have the compiler check everything, and then have the linker pull in
what is actually needed.

Even without further configure knobs, this reduces the size of the
loadable segments of the rauc binary from 224k to 181k for a
native (x86-64) build with --disable-service - obviously exact numbers
will vary with architecture and compiler.

The generated rauc.map file can be used to get an idea of what
actually gets discarded. Unsurprisingly, for --disable-service, most
of service.o is not included in the final link (since it is only
referenced from r_service_run(), whose reference from main.c is
compiled out for !ENABLE_SERVICE). But

$ size src/.libs/librauc_la-service.o
   text    data     bss     dec     hex filename
   6080       0      20    6100    17d4 src/.libs/librauc_la-service.o

so that cannot account for 43k. It turns out that a good deal of
rauc-installer-generated.o also gets discarded.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>