configure: allow disabling bundle, resign and convert commands
These commands are not generally useful on the embedded target, so it
can be useful to be able to remove all the associated code from the
generated binary. Building with --disable-create on x86-64 cuts the
size of the the loadable segments from 181k to 162k, i.e. about 10% -
results will obviously vary with architecture and compiler, but it's a
reasonable ballpark figure.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>