src/update_handler: provide RAUC_BUNDLE_MOUNT_POINT to hook script
A pre- or post-hook is often used to install extra artifacts from the
bundle, so it makes sense to provide the path where the bundle is
mounted, just as it is done for global install handlers.

I _think_ this might always be the same as
${RAUC_MOUNT_PREFIX}/bundle, but I can't find that explicitly stated,
and I think it's better to have the actual value provided directly
instead of having to know a magic "/bundle" suffix.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>