src/install: replace RAUC_UPDATE_SOURCE by RAUC_BUNDLE_MOUNT_POINT
A subsequent patch will expose the mount point of the bundle also to
hooks, but it has been agreed that RAUC_BUNDLE_MOUNT_POINT is a better
name to use for this (it also matches the RAUC_SLOT_MOUNT_POINT
exposed to the hooks). So update install.c and the documentation to
use that new name, but keep RAUC_UPDATE_SOURCE for compatibility with
existing handlers.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>