src/main: replace ifdeffery with G_GNUC_UNUSED macro
The code has #ifdeffery to avoid compiling in the service_start()
function when ENABLE_SERVICE is not set. I assume that's mostly for
avoiding a warning about an unused function. However, for embedded
targets, it's also important to reduce code size as much as possible,
and even with ENABLE_SERVICE==0, r_service_run() is still compiled in,
despite never being called.

We'll later add compiler and linker flags to eliminate such unused
functions, and also add configure options to allow disabling even more
rauc subcommands (e.g. being able to sign or resign a bundle is
generally not needed on embedded targets). To avoid having to #ifdef
too much, and to ensure we always get compile checking of as much code
as possible, start using the G_GNUC_UNUSED macro (aka the gcc
__attribute__((__unused__))) to annotate such potentially unreferenced
functions.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>