src: add option to enable Certificate Revocation List (CRL) checking
In order to actually use a provided CRL for checking, some flags need to
be set in OpenSSL. These were unset up to now. As a result of this, a
keyring / cert store file or directory could have contained a CRL, but
the Bundle verification process in RAUC simply ignored it.

This patch now introduces the [keyring] section option 'check-crl' that
allows enabling CRL handling for RAUC's signature verification.

The reason why this cannot be enabled by default is that OpenSSL would
enforce having a CRL in a cert store. This will not only break existing
setups but also make the simplest case (a simple self-signed
certificate for now) more complex.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>