configure.ac: only check for systemd and dbus dirs if the service is enabled
Installing the dbus files without the systemd unit is hardly sensible,
so let the --enable-service switch cover them all.

Further this allows to determine the dbus and systemd paths only if the
service should be enabled.

Note this drops the parameter --with-systemdunitdir of configure. Use

	./configure SYSTEMD_SYSTEMUNITDIR=...

instead. Also this adds error handling if the systemd pkg-config query
fails.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>