src/config_file.c: support explicit selection of per-slot status file
This aims to move further away from having the per-slot status file as a
default. With the option to explicitly select per-slot status storage,
we can now always note in the log that the integrator did not choose a
status storage strategy explicitly.

There are valid use-cases where the local slot status should be
preferred, e.g. when there is no (non-redundant) central storage location
available. Thus allowing an explicit choice is a better method than
noting that only a local slot status is used.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>