src/bundle: add sanity check for having CRLs but check-crl not set
We cannot enable CRLs by default but as stated in the respective commit.
Nevertheless, there may be setups that probably already have a CRL in
their key store.
To more visibly tell them to enable the 'check-crl' flag, we check
stores for files that have the standardized CRL header in it. If these
are present and check-crl is not enabled, we print a warning to inform
the user about the ineffectiveness of their CRLs.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>