test/config_file: fix system-serial test
While this test is not very effective anyway, its functionality was only
based on a chain of incorrect implementations around context handling.

It does not initialize a context but uses the context from the previous
builds. This only points to a matching memory locations as there was
neither a cleanup for the respective variable in the context freeing
functions nor was the context freeing called on context test tear down.

We fix this test by simply also calling the common context test set up
and tear down functions.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>