treewide: replace erroneous uses of g_return_if_fail for flow control
On early exit, cleanup functions may be called with a NULL argument.
This is completely normal behavior and not a programming error
and thus should not be guarded against with g_return_if_fail.

These g_return_* functions expand to nothing if G_DISABLE_CHECKS and
to a panic when G_DEBUG=fatal-warnings is defined, both of which
is not expected behavior.

All cleanup functions we have are affected, so use a normal if condition
in all of them.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>