src/update_handler: set RAUC_SLOT_BOOTNAME for child slots too
The RAUC_SLOT_BOOTNAME environment variable was passed to hooks for the
parent slot in a group, but not the child slots. There is no easy way
for a child slot hook to figure out what the bootname would be for the
parent hook in its group.

With this change, RAUC_SLOT_BOOTNAME will be set consistently for all
slots in a group.

Signed-off-by: Dan Callaghan <dan.callaghan@opengear.com>