test/config_file: enhance debugability load_config() call failures in tests
This replaces simple gboolean return validation of load_config() by first
checking the error with g_assert_no_error() and then validating the return
value previously stored in a variable.

In case of test failures this allows having the error message in the
log as we do not abort prior to evaluating it.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>