src: require config loading to succeed if config path is configured
If one specified a system config, we expect this to be used instead of
loading useless defaults on errors.

Only if no system config is given, explicitly and silently use a default
config.

Latter is required as we cannot simply get rid off the dependencies in
the code on having a config structure set up at all.

The service_test_install_api() test must also be marked as skipped now
as fixture_helper_set_up_system() aborts anyway when not running as root
and thus does not set up a valid system.conf

Due to enforced system.conf loading, this would now make service start
fail instead of starting with useless defaults.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>