.travis.yml: support testing non-service build too
So far, RAUC supports being built with or without servie enabled.
This has a lot of impact on the resulting binaries that are either built
with a D-Bus API or direct method calls.

Thus we should also test the non-service case in CI too which requires
RAUC to be configured differently.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>