src/config_file: fix checking for unknown keys in slot configuration
The intention of removing used keys is to allow checking that no unknown
key is remaining and issue an error in favor of not interpreting e.g.
misspelled keys.

However, due to the split-up of slot parsing for parent handling
combined with an early 'continue' statement, the actual checking of
remaining keys did not happen for most slots.

To simplify both the parent lookup as well as the config checking, this
patch introduces a new internally used member of RaucSlot structure,
called 'parent_name'. This allows storing the original parsed parent
name and thus enables us to cleanly do the config file parsing (and
remaining keys handling) in the first loop and the parent resolving in
a second independent one that simply uses the string value stored in
'parent_name'.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>